fix: MID4-93 최근 작성 댓글 좋아요 수 집계 반영

+++ src/main/java/com/codeit/sb13/monew/comment/repository/dto/RecentCommentActivityProjection.java
@@ -10,7 +10,7 @@ public record RecentCommentActivityProjection(
         UUID userId,
         String userNickname,
         String content,
-        Integer likeCount,
+        Long likeCount,
         LocalDateTime createdAt
 ) {
 }
